Move APIVersions out of models
