+++ catalogue_api/api/src/main/scala/uk/ac/wellcome/platform/api/controllers/SwaggerController.scala
@@ -6,7 +6,7 @@ import com.twitter.finatra.http.Controller
 import io.swagger.util.Json
 import io.swagger.models.{Info, Scheme, Swagger}
 import javax.inject.{Inject, Singleton}
-import uk.ac.wellcome.models.ApiVersions
+import uk.ac.wellcome.versions.ApiVersions
 
 object ApiV1Swagger extends Swagger
 object ApiV2Swagger extends Swagger
