+++ common/src/main/scala/uk/ac/wellcome/versions/ApiVersions.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models
+package uk.ac.wellcome.versions
 
 object ApiVersions extends Enumeration {
   val v1, v2 = Value
